fix(flutter): resolve dependency conflict + add missing API surfaces
1. Pod install gagal di Codemagic karena conflict:
   - mobile_scanner 5.x → GoogleMLKit/MLKitCommon 11.x → GoogleDataTransport <10.0
   - Firebase 11.x → GoogleDataTransport ~> 10.0
   Solusi: disable mobile_scanner dari pubspec sementara (tidak dipakai
   di lib/screens/), stub barcode_scanner_modal.dart jadi no-op.

2. Banyak file user existing di flutter_app/ expect API surface yang
   belum ada di foundation stubs. Tambah:
   - ApiConfig.baseUrl alias (selain apiBaseUrl)
   - ReadOnlyMode.isReadOnly + assertWritable aliases
   - ApiClient.clearSession + putJson + patchJson + postMultipartFile
     (dengan filename + contentType params)
   - MemberProfile.fromApiJson factory (auto-unwrap {user: ...})
   - MemberAddress model (untuk shipping_service)
   - MemberVoucher model (untuk voucher_service)
   - models/member_profile.dart re-export sister files
   - FeedAuthor.role param
   - MotionPrefs.effective(context, duration) helper
   - AppCrashlytics.setCustomKey method
   - NataloColors: textMuted, divider, feedDark, feedTextMuted, nataloBlue aliases
   - NataloTextStyles class (productPrice, productTitle, caption)
   - formatters: formatDateTime alias
   - BottomNavBar: onDestinationSelected callback param

3. Auth flow + cart price type fixes:
   - auth_service.login: null-check fetchProfile result
   - cart_service: cast price double → int (Product.price = int)

4. Codemagic.yaml debug: tambah diagnostic step + verbose flag supaya
   Apple API error muncul lengkap saat signing fail.

flutter analyze: 0 error (was 89 errors pasca git state corruption +
file merge dari natalopetshopflutter clone).

+++ flutter_app/lib/services/cart_service.dart
@@ -10,7 +10,10 @@ class CartService {
     if (rawItems is! List) return [];
 
     return rawItems.whereType<Map<String, dynamic>>().map((item) {
-      final price = _asDouble(item['price']);
+      // Product.price = int (rupiah). Cart endpoint kadang return double
+      // (legacy formatting). Truncate ke int aman karena harga rupiah
+      // tidak punya desimal.
+      final price = _asDouble(item['price']).toInt();
       final product = Product(
         id: (item['productId'] ?? '').toString(),
         slug: (item['productId'] ?? '').toString(),
